# src/swagger.yaml
openapi: 3.0.0
info:
  title: TaskTrackr API
  version: 1.0.0
  description: API for TaskTrackr To-Do List App
servers:
  - url: http://localhost:5000/api
paths:
  /auth/register:
    post:
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string }
                password: { type: string }
                role: { type: string, enum: ['USER', 'ADMIN'] }
      responses:
        '201': { description: User registered }
        '400': { description: Bad request }
  /auth/login:
    post:
      summary: Login and get JWT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string }
                password: { type: string }
      responses:
        '200': { description: JWT token }
        '401': { description: Invalid credentials }
  /auth/token:
    post:
      summary: Generate JWT (demo)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string }
                role: { type: string }
      responses:
        '200': { description: JWT token }
        '404': { description: User not found }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT